requirements.txt
pandas
numpy
tensorflow
scikit-learn
seaborn
matplotlib
pyyaml